# frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Your API Title",
    "description": "Your API Description",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/player/update": {
      "put": {
        "tags": [
          "Player"
        ],
        "operationId": "updatePlayer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/questions/validate-answer": {
      "post": {
        "tags": [
          "Questions"
        ],
        "operationId": "validateAnswer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerValidationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/player/{playerId}/profile-picture-upload": {
      "post": {
        "tags": [
          "Player"
        ],
        "operationId": "uploadProfilePicture",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/notification/send/{playerId}": {
      "post": {
        "tags": [
          "Notification"
        ],
        "operationId": "sendNotification",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/multiplayer/{sessionId}/complete/{playerId}": {
      "post": {
        "tags": [
          "Multiplayer"
        ],
        "operationId": "giveUp",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/multiplayer/session/{sessionId}/rematch-request/{playerId}": {
      "post": {
        "tags": [
          "Multiplayer"
        ],
        "operationId": "requestRematch",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/multiplayer/session/{sessionId}/complete/{playerId}": {
      "post": {
        "tags": [
          "Multiplayer"
        ],
        "operationId": "acknowledgeGameOver",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/friendship/block": {
      "post": {
        "tags": [
          "Friendship"
        ],
        "operationId": "blockPlayer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerInteraction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/friendship/add": {
      "post": {
        "tags": [
          "Friendship"
        ],
        "operationId": "addFriend",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerInteraction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/questions/{sessionId}/category-selection": {
      "get": {
        "tags": [
          "Questions"
        ],
        "operationId": "getThreeRandomCategories",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/questions/session/{sessionId}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "operationId": "getSessionQuestions",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/questions/session/current-questions": {
      "get": {
        "tags": [
          "Questions"
        ],
        "operationId": "getCurrentQuestions",
        "parameters": [
          {
            "name": "currentQuestionIds",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/questions/random-questions": {
      "get": {
        "tags": [
          "Questions"
        ],
        "operationId": "getThreeQuestionsByCategory",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/player/{playerId}": {
      "get": {
        "tags": [
          "Player"
        ],
        "operationId": "getPlayerById",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDTO"
                }
              }
            }
          }
        }
      }
    },
    "/notification/{playerId}": {
      "get": {
        "tags": [
          "Notification"
        ],
        "operationId": "getPlayerNotifications",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/multiplayer/{playerId}": {
      "get": {
        "tags": [
          "Multiplayer"
        ],
        "operationId": "getPlayerSessions",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MultiplayerSessionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/multiplayer/session/{sessionId}": {
      "get": {
        "tags": [
          "Multiplayer"
        ],
        "operationId": "getGameState",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/friendship/friends/{playerId}": {
      "get": {
        "tags": [
          "Friendship"
        ],
        "operationId": "getFriends",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/multiplayer/matchmaking/cancel/{playerId}": {
      "delete": {
        "tags": [
          "Multiplayer"
        ],
        "operationId": "cancelMatchmaking",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UpdatePlayerRequest": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "integer",
            "format": "int64"
          },
          "updateField": {
            "type": "string",
            "enum": [
              "DISPLAY_NAME"
            ]
          },
          "newDisplayName": {
            "type": "string"
          }
        }
      },
      "AnswerValidationRequest": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "sessionId": {
            "type": "integer",
            "format": "int64"
          },
          "answer": {
            "type": "string"
          },
          "playerId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AnswerValidationResponse": {
        "type": "object",
        "properties": {
          "correctAnswerIndex": {
            "type": "integer",
            "format": "int32"
          },
          "correct": {
            "type": "boolean"
          }
        }
      },
      "PlayerInteraction": {
        "type": "object",
        "properties": {
          "senderId": {
            "type": "integer",
            "format": "int64"
          },
          "receiverUsername": {
            "type": "string"
          }
        }
      },
      "RegistrationRequest": {
        "required": [
          "displayName",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "AuthRequest": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "playerDTO": {
            "$ref": "#/components/schemas/PlayerDTO"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "CategoryStats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PlayerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/Stats"
          },
          "base64Image": {
            "type": "string"
          }
        }
      },
      "Stats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "numOfGames": {
            "type": "integer",
            "format": "int32"
          },
          "numOfWins": {
            "type": "integer",
            "format": "int32"
          },
          "numOfLosses": {
            "type": "integer",
            "format": "int32"
          },
          "categoryStats": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/CategoryStats"
            }
          }
        }
      },
      "QuestionDTO": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "question": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "playerId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": [
              "FRIEND_REQUEST",
              "REMATCH_REQUEST",
              "REMATCH_ACCEPTED",
              "GAME_WON",
              "GAME_LOST"
            ]
          },
          "pendingSessionId": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "read": {
            "type": "boolean"
          }
        }
      },
      "MultiplayerSessionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "playerDTOList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerDTO"
            }
          },
          "score": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "COMPLETED"
            ]
          },
          "currentPlayerTurn": {
            "$ref": "#/components/schemas/PlayerDTO"
          }
        }
      },
      "GameStateResponse": {
        "type": "object",
        "properties": {
          "playerTurn": {
            "type": "integer",
            "format": "int64"
          },
          "playerDTOS": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerDTO"
            }
          },
          "currentQuestionIndex": {
            "type": "integer",
            "format": "int32"
          },
          "scores": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "COMPLETED"
            ]
          },
          "results": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerQuestionResult"
            }
          },
          "questionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "roundCategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "playerAcknowledgment": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        }
      },
      "PlayerQuestionResult": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "integer",
            "format": "int64"
          },
          "questionId": {
            "type": "integer",
            "format": "int64"
          },
          "questionIndex": {
            "type": "integer",
            "format": "int32"
          },
          "correct": {
            "type": "boolean"
          }
        }
      }
    }
  }
}